Create Any Base Multiplication
If n1 = 29 and n2 = 8, swap them so that n1 becomes bottom number. i.e. n2 = 29 and n1 = 8, OR
29    <--- Top Number
x8    <---Bottom Number
____
____

We did that because in our program, we have considered n1 to be at the bottom for which n1 <= n2. You can take any of n1, n2 as the bottom number.